chore: rename package and file apiimpl/utils to observability
