+++ comp/api/api/apiimpl/server_cmd.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/comp/api/api/apiimpl/internal/agent"
 	"github.com/DataDog/datadog-agent/comp/api/api/apiimpl/internal/check"
-	apiutils "github.com/DataDog/datadog-agent/comp/api/api/apiimpl/utils"
+	"github.com/DataDog/datadog-agent/comp/api/api/apiimpl/observability"
 	api "github.com/DataDog/datadog-agent/comp/api/api/def"
 	"github.com/DataDog/datadog-agent/comp/collector/collector"
 	"github.com/DataDog/datadog-agent/comp/core/autodiscovery"
@@ -143,7 +143,7 @@ func startCMDServer(
 	cmdMux.Handle("/", gwmux)
 
 	// Add some observability in the API server
-	cmdMuxHandler := apiutils.LogResponseHandler(cmdServerName)(cmdMux)
+	cmdMuxHandler := observability.LogResponseHandler(cmdServerName)(cmdMux)
 
 	srv := grpcutil.NewMuxedGRPCServer(
 		cmdAddr,
