+++ comp/api/api/apiimpl/server_ipc.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	configendpoint "github.com/DataDog/datadog-agent/comp/api/api/apiimpl/internal/config"
-	apiutils "github.com/DataDog/datadog-agent/comp/api/api/apiimpl/utils"
+	"github.com/DataDog/datadog-agent/comp/api/api/apiimpl/observability"
 	"github.com/DataDog/datadog-agent/pkg/config"
 )
 
@@ -33,7 +33,7 @@ func startIPCServer(ipcServerAddr string, tlsConfig *tls.Config) (err error) {
 	ipcMux.Handle(
 		"/config/v1/",
 		http.StripPrefix("/config/v1", configEndpointMux))
-	ipcMuxHandler := apiutils.LogResponseHandler(ipcServerName)(ipcMux)
+	ipcMuxHandler := observability.LogResponseHandler(ipcServerName)(ipcMux)
 
 	ipcServer := &http.Server{
 		Addr:      ipcServerAddr,
